+++ CHANGELOG.md
@@ -25,6 +25,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 - Option to add list of rows to table and list of cells to table row in their respective builders
 - Support for document object pdf tagging/accessibility metadata
 - Added Hierarchy task and its accompanying html visualisation
+- External display rule support along with metadata, subject, target aliasing,
+  base template override, variable structure override and target folder
+  override support
 
 ### Changed